requirements.txt
pytest
pytest-mock
pre-commit
flit
pydantic==0.32
pylint
nox
toml
pyyaml==5.3.1
mocker
